Tagging the 0.12.7 release of QStarDict

+++ qstardict/speaker.cpp
@@ -41,6 +41,9 @@ Speaker::~Speaker()
 
 void Speaker::speak(const QString &word)
 {
+    if (m_speechCmd.isEmpty())
+        return;
+
     if (m_speechProcess->state() != QProcess::NotRunning)
     m_speechProcess->kill();
